erofs-utils: implement 48-bit block addressing for unencoded inodes
.. Also port the same kernel logic to erofsfuse and dump.erofs.

Link: https://lore.kernel.org/r/20250704074535.2308212-6-hsiangkao@linux.alibaba.com

+++ lib/blobchunk.c
@@ -157,7 +157,7 @@ int erofs_blob_write_chunk_indexes(struct erofs_inode *inode,
 		chunk = *(void **)(inode->chunkindexes + src);
 
 		if (chunk->blkaddr == EROFS_NULL_ADDR) {
-			idx.startblk_lo = EROFS_NULL_ADDR;
+			idx.startblk_lo = (u32)EROFS_NULL_ADDR;
 		} else if (chunk->device_id) {
 			DBG_BUGON(!(inode->u.chunkformat & EROFS_CHUNK_FORMAT_INDEXES));
 			idx.startblk_lo = chunk->blkaddr;
